got NV_DX_interop working

+++ native2/lib/DriftGL/src/DriftGL_glx.cpp
@@ -151,7 +151,7 @@ namespace driftgl {
 			return ctx;
 		}
 
-	Context* WrapContext(long nativeContextHandle) {
+	Context* WrapContext(long long nativeContextHandle) {
 		DriftGLXContext* ctx = new DriftGLXContext();
 		ctx->glxContext = (GLXContext) nativeContextHandle;
 		return ctx;
